feat: utiliza do lerna init
