+++ lerna.json
@@ -1,0 +1,5 @@
+{
+  "packages": ["packages/*"],
+  "version": "1.1.0-dev.16",
+  "npmClients": "npm"
+}
