+++ package.json
@@ -1,12 +1,12 @@
 {
   "name": "@animaliads/web-components",
-  "version": "1.1.0-dev.14",
+  "version": "1.1.0-dev.16",
   "description": "Repositório de Web Components do Animalia DS",
-  "main": "src/index.js",
+  "main": "index.js",
   "tag": "dev",
   "type": "module",
   "scripts": {
-    "build": "rimraf dist && rollup -c",
+    "build": "rimraf dist && lerna exec -- rollup -c ../../scripts/rollup.config.js",
     "dev": "rollup -c -w",
     "start": "live-server --open=demo",
     "test": "jest --coverage",
@@ -39,52 +39,52 @@
   },
   "homepage": "https://github.com/animaliads/animalia-web-components#readme",
   "devDependencies": {
-    "@babel/core": "^7.12.10",
-    "@babel/plugin-syntax-class-properties": "^7.12.1",
-    "@babel/plugin-transform-classes": "^7.12.1",
-    "@babel/preset-env": "^7.12.11",
-    "@babel/preset-typescript": "^7.12.7",
-    "@commitlint/cli": "^11.0.0",
-    "@commitlint/config-angular": "^11.0.0",
+    "@babel/core": "^7.15.0",
+    "@babel/plugin-syntax-class-properties": "^7.12.13",
+    "@babel/plugin-transform-classes": "^7.14.9",
+    "@babel/preset-env": "^7.15.0",
+    "@babel/preset-typescript": "^7.15.0",
+    "@commitlint/cli": "^13.1.0",
+    "@commitlint/config-angular": "^13.1.0",
     "@lerna/batch-packages": "^3.16.0",
-    "@lerna/filter-packages": "^3.18.0",
-    "@lerna/project": "^3.21.0",
-    "@rollup/plugin-babel": "^5.2.2",
-    "@rollup/plugin-node-resolve": "^11.2.1",
-    "@rollup/plugin-typescript": "^8.1.0",
-    "@storybook/addon-a11y": "^6.1.20",
-    "@storybook/addon-docs": "^6.1.20",
-    "@storybook/addon-essentials": "^6.1.20",
-    "@storybook/addon-links": "^6.1.20",
-    "@storybook/addons": "^6.1.20",
-    "@storybook/web-components": "^6.1.20",
-    "@types/jest": "^26.0.20",
-    "@typescript-eslint/eslint-plugin": "^4.11.0",
-    "@typescript-eslint/parser": "^4.11.0",
-    "@webcomponents/webcomponentsjs": "^2.5.0",
-    "chromatic": "^5.6.3",
-    "concurrently": "^5.3.0",
-    "eslint": "^7.16.0",
-    "eslint-config-prettier": "^7.1.0",
-    "eslint-plugin-unused-imports": "^1.0.1",
-    "husky": "^4.3.6",
-    "jest": "^26.6.3",
+    "@lerna/filter-packages": "^4.0.0",
+    "@lerna/project": "^4.0.0",
+    "@rollup/plugin-babel": "^5.3.0",
+    "@rollup/plugin-node-resolve": "^13.0.4",
+    "@rollup/plugin-typescript": "^8.2.5",
+    "@storybook/addon-a11y": "^6.3.7",
+    "@storybook/addon-docs": "^6.3.7",
+    "@storybook/addon-essentials": "^6.3.7",
+    "@storybook/addon-links": "^6.3.7",
+    "@storybook/addons": "^6.3.7",
+    "@storybook/web-components": "^6.3.7",
+    "@types/jest": "^27.0.1",
+    "@typescript-eslint/eslint-plugin": "^4.30.0",
+    "@typescript-eslint/parser": "^4.30.0",
+    "@webcomponents/webcomponentsjs": "^2.6.0",
+    "chromatic": "^5.9.2",
+    "concurrently": "^6.2.1",
+    "eslint": "^7.32.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-unused-imports": "^1.1.4",
+    "husky": "^7.0.2",
+    "jest": "^27.1.0",
     "lerna": "^4.0.0",
-    "lit-element": "^2.4.0",
-    "lit-html": "^1.3.0",
+    "lit-element": "^2.5.1",
+    "lit-html": "^1.4.1",
     "live-server": "^1.2.1",
     "minimist": "^1.2.5",
-    "prettier": "2.2.1",
-    "pretty-quick": "^3.1.0",
-    "rollup": "^2.51.1",
-    "rollup-plugin-copy": "^3.3.0",
-    "rollup-plugin-minify-html-literals": "^1.2.5",
+    "prettier": "2.3.2",
+    "pretty-quick": "^3.1.1",
+    "rollup": "^2.56.3",
+    "rollup-plugin-copy": "^3.4.0",
+    "rollup-plugin-minify-html-literals": "^1.2.6",
     "rollup-plugin-terser": "^7.0.2",
     "storybook-addon-pseudo-states": "^1.0.0",
-    "ts-jest": "^26.4.4",
-    "ts-loader": "^8.0.17",
-    "ts-node": "^9.1.1",
-    "tslib": "^2.0.3",
-    "typescript": "^4.1.3"
+    "ts-jest": "^27.0.5",
+    "ts-loader": "^9.2.5",
+    "ts-node": "^10.2.1",
+    "tslib": "^2.3.1",
+    "typescript": "4.3.5"
   }
 }
