--- packages/animalia-web-components/package.json
@@ -1,31 +1,0 @@
-{
-  "name": "@animaliads/web-components",
-  "version": "0.0.0-PLACEHOLDER",
-  "description": "Web Components do Animalia DS",
-  "tag": "dev",
-  "keywords": [
-    "design system",
-    "javascript",
-    "web components",
-    "button",
-    "link"
-  ],
-  "main": "animalia-web-components/index.js",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/animaliads/animalia-web-components.git",
-    "directory": "packages/animalia-web-components"
-  },
-  "license": "MIT",
-  "dependencies": {
-    "@animaliads/animalia-brand": "^1.0.0-dev.1",
-    "@animaliads/css-tokens": "^1.0.0-dev.1",
-    "tslib": "^2.0.3"
-  },
-  "devDependencies": {
-    "eslint": "^7.16.0"
-  },
-  "publishConfig": {
-    "access": "public"
-  }
-}
