--- packages/animalia-web-components/src/index.ts
@@ -1,4 +1,0 @@
-export * from '../../button/src/index';
-export * from '../../link/src/index';
-export * from '../../radio/src/index';
-export * from '../../checkbox/src/index';
