--- packages/animalia-web-components/tsconfig.json
@@ -1,7 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "outDir": "."
-  },
-  "include": ["./src"]
-}
